requirements.txt
requests==2.32.4
apache-airflow==3.0.0
apache-airflow-providers-postgres==6.1.3